tests/wpt/web-platform-tests/css/css-text/parsing/text-transform-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-transform with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-transform">
<meta name="assert" content="text-transform supports the full grammar 'none | [capitalize | uppercase | lowercase ] || full-width || full-size-kana'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-transform", "none");

test_valid_value("text-transform", "capitalize");
test_valid_value("text-transform", "uppercase");
test_valid_value("text-transform", "lowercase");
test_valid_value("text-transform", "full-width");
test_valid_value("text-transform", "full-size-kana");

test_valid_value("text-transform", "capitalize full-width");
test_valid_value("text-transform", "uppercase full-size-kana");
test_valid_value("text-transform", "full-width full-size-kana");
test_valid_value("text-transform", "full-width lowercase");
test_valid_value("text-transform", "full-size-kana capitalize");
test_valid_value("text-transform", "full-size-kana full-width");

test_valid_value("text-transform", "capitalize full-width full-size-kana");
test_valid_value("text-transform", "full-width full-size-kana uppercase");
test_valid_value("text-transform", "full-size-kana lowercase full-width");
test_valid_value("text-transform", "lowercase full-size-kana full-width");
test_valid_value("text-transform", "full-width uppercase full-size-kana");
test_valid_value("text-transform", "full-size-kana full-width capitalize");
</script>
</body>
</html>
