third_party/blink/web_tests/external/wpt/css/css-text/parsing/text-transform-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['text-transform'] = "full-width" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-size-kana" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "capitalize full-width" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "uppercase full-size-kana" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-width full-size-kana" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-width lowercase" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-size-kana capitalize" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-size-kana full-width" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "capitalize full-width full-size-kana" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-width full-size-kana uppercase" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-size-kana lowercase full-width" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "lowercase full-size-kana full-width" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-width uppercase full-size-kana" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-transform'] = "full-size-kana full-width capitalize" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
